Refactoring the TestRiskLevel fragment to demo our new viewmodel approach. Part2
