+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/util/viewmodel/VDCSource.kt
@@ -15,7 +15,6 @@ class VDCSource @AssistedInject constructor(
     @Assisted private val assistAction: ((VDCFactory<out VDC>, SavedStateHandle) -> VDC)?
 ) : AbstractSavedStateViewModelFactory(savedStateOwner, defaultSavedState) {
 
-
     override fun <T : ViewModel?> create(
         key: String,
         modelClass: Class<T>,
